+++ build.clj
@@ -8,7 +8,7 @@
    [org.corfield.build :as bb]))
 
 (def lib 'io.staticweb/cloudformation-templating)
-(def version "3.0.0")
+(def version "3.1.0")
 
 (defn get-version [opts]
   (str version (when (:snapshot opts) "-SNAPSHOT")))